Fix arrays getting messed up
Man this was an  annoying bug. the resolveOptions function was changing
arrays into objects with "0", "1", etc as keys. Which made them not
iterable. We thought it was firebase's fault, fixed it there. The issue still
persisted. I think it took us about an hour to find this function as the
cause.